small refactoring and add profanity stop
