+++ src/components/User/ChangeProfile.js
@@ -1,7 +1,7 @@
 import FormSubmitButton from "common/FormSubmitButton";
 import InputFormRow from "common/InputFormRow";
 import { USER_PROFILE } from "common/fields";
-import { hostUrl } from '../../common/urls';
+import { hostUrl } from '../../utils/urls';
 
 export default function ChangeProfile({ handleOnChange, userInfo, user }) {
     const updateUserInfo = () => {
